Move added EventHandler to its own file
