+++ src/Microsoft.Fast.Components.FluentUI/BindAttributes.cs
@@ -1,5 +1,4 @@
-﻿using System;
-
+﻿
 using Microsoft.AspNetCore.Components;
 
 namespace Microsoft.Fast.Components.FluentUI
@@ -56,16 +55,6 @@ namespace Microsoft.Fast.Components.FluentUI
 
 
     public static class BindAttributes
-    {
-    }
-
-    public class CheckboxChangeEventArgs : EventArgs
-    {
-        public bool Checked { get; set; }
-    }
-
-    [EventHandler("onfluentcheckedchange", typeof(CheckboxChangeEventArgs), enableStopPropagation: true, enablePreventDefault: true)]
-    public static class EventHandlers
     {
     }
 }
